scripts, tests, rpc: add test for system_networkState (#783)
