+++ tests/rpc/common.go
@@ -1,0 +1,59 @@
+// Copyright 2019 ChainSafe Systems (ON) Corp.
+// This file is part of gossamer.
+//
+// The gossamer library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The gossamer library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
+
+package rpc
+
+import (
+	"encoding/json"
+	"os"
+	"time"
+)
+
+//nolint
+var (
+	GOSSAMER_INTEGRATION_TEST_MODE = os.Getenv("GOSSAMER_INTEGRATION_TEST_MODE")
+
+	GOSSAMER_NODE_HOST = os.Getenv("GOSSAMER_NODE_HOST")
+
+	ContentTypeJSON   = "application/json"
+	dialTimeout       = 60 * time.Second
+	httpClientTimeout = 120 * time.Second
+)
+
+type serverResponse struct {
+	// JSON-RPC Version
+	Version string `json:"jsonrpc"`
+	// Resulting values
+	Result json.RawMessage `json:"result"`
+	// Any generated errors
+	Error *Error `json:"error"`
+	// Request id
+	ID *json.RawMessage `json:"id"`
+}
+
+// ErrCode is a int type used for the rpc error codes
+type ErrCode int
+
+// Error is a struct that holds the error message and the error code for a error
+type Error struct {
+	Message   string
+	ErrorCode ErrCode
+}
+
+// Error returns the error Message string
+func (e *Error) Error() string {
+	return e.Message
+}
